refactor: 유효성 검증 함수들 validators.js로 추출
