+++ src/controller/controller.js
@@ -1,17 +1,7 @@
-import {
-  $,
-  isBlank,
-  isIncludeSpace,
-  isDuplicated,
-  splitUsingComma,
-  isOverFiveLetters,
-  hasEmptyElement,
-  generateCars,
-  isZero,
-  hasSpecial,
-} from '../utils/utils.js';
+import { $, generateCars } from '../utils/utils.js';
 import { default as UI } from '../view/view.js';
-import { ERROR, SELECTOR } from '../utils/constants.js';
+import { SELECTOR } from '../utils/constants.js';
+import { isValidCarNames, isValidRacingCount } from '../utils/validators.js';
 
 export default function RacingCarGame() {
   const init = () => {
@@ -33,24 +23,10 @@ export default function RacingCarGame() {
     UI.showRacingCountSection();
   };
 
-  const isValidCarNames = string => {
-    if (isBlank(string)) return alert(ERROR.BLANK_SUBMIT);
-    if (isIncludeSpace(string)) return alert(ERROR.INCLUDE_SPACE);
-
-    const carNamesArray = splitUsingComma(string);
-    if (isDuplicated(carNamesArray)) return alert(ERROR.DUPLICATED);
-    if (isOverFiveLetters(carNamesArray)) return alert(ERROR.OVER_FIVE_LETTERS);
-    if (hasEmptyElement(carNamesArray)) return alert(ERROR.HAS_EMPTY_ELEMENT);
-
-    return true;
-  };
-
   const handleRacingCountSubmit = e => {
     e.preventDefault();
 
-    if (isBlank(UI.getRacingCount())) return alert(ERROR.BLANK_SUBMIT);
-    if (isZero(UI.getRacingCount())) return alert(ERROR.NOT_POSIVITE_INT);
-    if (hasSpecial(UI.getRacingCount())) return alert(ERROR.INCLUDE_SPECIAL);
+    if (!isValidRacingCount(UI.getRacingCount())) return;
 
     playCarRacing();
